refactor: moved `LoadingErrorComponent`
