+++ ts/components/LoadingErrorComponent.tsx
@@ -1,12 +1,12 @@
 import { Millisecond } from "@pagopa/ts-commons/lib/units";
 import * as React from "react";
 import { useEffect } from "react";
-import { View, ActivityIndicator, SafeAreaView } from "react-native";
-import { IOStyles } from "../../../../../components/core/variables/IOStyles";
-import { InfoScreenComponent } from "../../../../../components/infoScreen/InfoScreenComponent";
-import GenericErrorComponent from "../../../../../components/screens/GenericErrorComponent";
-import { WithTestID } from "../../../../../types/WithTestID";
-import { setAccessibilityFocus } from "../../../../../utils/accessibility";
+import { ActivityIndicator, SafeAreaView, View } from "react-native";
+import { WithTestID } from "../types/WithTestID";
+import { setAccessibilityFocus } from "../utils/accessibility";
+import { IOStyles } from "./core/variables/IOStyles";
+import { InfoScreenComponent } from "./infoScreen/InfoScreenComponent";
+import GenericErrorComponent from "./screens/GenericErrorComponent";
 
 export type LoadingErrorProps = WithTestID<{
   isLoading: boolean;
