+++ ts/features/bonus/bonusVacanze/screens/activation/LoadActivateBonusScreen.tsx
@@ -5,7 +5,7 @@ import I18n from "../../../../../i18n";
 import { GlobalState } from "../../../../../store/reducers/types";
 import { abortBonusRequest } from "../../components/alert/AbortBonusRequest";
 import { useHardwareBackButton } from "../../../../../hooks/useHardwareBackButton";
-import { LoadingErrorComponent } from "../../components/loadingErrorScreen/LoadingErrorComponent";
+import { LoadingErrorComponent } from "../../../../../components/LoadingErrorComponent";
 import {
   activateBonusVacanze,
   cancelBonusVacanzeRequest
