+++ ts/features/bonus/bpd/screens/iban/IbanLoadingUpsert.tsx
@@ -7,7 +7,7 @@ import { Iban } from "../../../../../../definitions/backend/Iban";
 import I18n from "../../../../../i18n";
 import { GlobalState } from "../../../../../store/reducers/types";
 import { useHardwareBackButton } from "../../../../../hooks/useHardwareBackButton";
-import { LoadingErrorComponent } from "../../../bonusVacanze/components/loadingErrorScreen/LoadingErrorComponent";
+import { LoadingErrorComponent } from "../../../../../components/LoadingErrorComponent";
 import {
   bpdIbanInsertionCancel,
   bpdUpsertIban
