+++ ts/features/bonus/bpd/screens/onboarding/LoadBpdActivationStatus.tsx
@@ -5,7 +5,7 @@ import { Dispatch } from "redux";
 import I18n from "../../../../../i18n";
 import { GlobalState } from "../../../../../store/reducers/types";
 import { useHardwareBackButton } from "../../../../../hooks/useHardwareBackButton";
-import { LoadingErrorComponent } from "../../../bonusVacanze/components/loadingErrorScreen/LoadingErrorComponent";
+import { LoadingErrorComponent } from "../../../../../components/LoadingErrorComponent";
 import {
   bpdOnboardingCancel,
   bpdOnboardingStart
