+++ ts/features/bonus/cdc/screens/CdcBonusRequestBonusRequested.tsx
@@ -10,7 +10,7 @@ import CdcRequirementsError from "../components/CdcRequirementsError";
 import { cdcEnrollUserToBonus } from "../store/actions/cdcBonusRequest";
 import { fold } from "../../../../common/model/RemoteValue";
 import CdcRequestCompleted from "../components/CdcRequestCompleted";
-import { LoadingErrorComponent } from "../../bonusVacanze/components/loadingErrorScreen/LoadingErrorComponent";
+import { LoadingErrorComponent } from "../../../../components/LoadingErrorComponent";
 import I18n from "../../../../i18n";
 import {
   CdcBonusRequestResponse,
