+++ ts/features/wallet/onboarding/cobadge/screens/search/LoadCoBadgeSearch.tsx
@@ -5,7 +5,7 @@ import I18n from "../../../../../../i18n";
 import { GlobalState } from "../../../../../../store/reducers/types";
 import { WithTestID } from "../../../../../../types/WithTestID";
 import { useHardwareBackButton } from "../../../../../../hooks/useHardwareBackButton";
-import { LoadingErrorComponent } from "../../../../../bonus/bonusVacanze/components/loadingErrorScreen/LoadingErrorComponent";
+import { LoadingErrorComponent } from "../../../../../../components/LoadingErrorComponent";
 import { searchUserCoBadge, walletAddCoBadgeCancel } from "../../store/actions";
 import { onboardingCoBadgeAbiSelectedSelector } from "../../store/reducers/abiSelected";
 import { onboardingCoBadgeFoundIsError } from "../../store/reducers/foundCoBadge";
