+++ ts/features/wallet/paypal/screen/PayPalPspUpdateScreen.tsx
@@ -33,7 +33,7 @@ import { useIOSelector } from "../../../../store/hooks";
 import { pspV2ListSelector } from "../../../../store/reducers/wallet/payment";
 import { emptyContextualHelp } from "../../../../utils/emptyContextualHelp";
 import { formatNumberCentsToAmount } from "../../../../utils/stringBuilder";
-import { LoadingErrorComponent } from "../../../bonus/bonusVacanze/components/loadingErrorScreen/LoadingErrorComponent";
+import { LoadingErrorComponent } from "../../../../components/LoadingErrorComponent";
 import { isError, isReady } from "../../../../common/model/RemoteValue";
 import { useImageResize } from "../../onboarding/bancomat/hooks/useImageResize";
 import {
