+++ ts/screens/wallet/payment/PickPspScreen.tsx
@@ -20,7 +20,7 @@ import FooterWithButtons from "../../../components/ui/FooterWithButtons";
 import { LightModalContextInterface } from "../../../components/ui/LightModal";
 import { PspComponent } from "../../../components/wallet/payment/PspComponent";
 import { cancelButtonProps } from "../../../components/buttons/ButtonConfigurations";
-import { LoadingErrorComponent } from "../../../features/bonus/bonusVacanze/components/loadingErrorScreen/LoadingErrorComponent";
+import { LoadingErrorComponent } from "../../../components/LoadingErrorComponent";
 import {
   getValueOrElse,
   isError,
